Move all test cases to switch file Order test cases to reflect specification document order Add comments for each test case from specification
